+++ packages/ai/src/registry/registry.ts
@@ -1,4 +1,5 @@
 import type { KnownProvider } from "@oh-my-pi/pi-catalog";
+import { agnesProvider } from "./agnes";
 import { aimlApiProvider } from "./aimlapi";
 import { alibabaCodingPlanProvider } from "./alibaba-coding-plan";
 import { alibabaTokenPlanProvider } from "./alibaba-token-plan";
@@ -145,6 +146,7 @@ const ALL = [
 	googleVertexProvider,
 	xaiProvider,
 	groqProvider,
+	agnesProvider,
 	mistralProvider,
 	minimaxProvider,
 	amazonBedrockProvider,